+++ app/frontend/Article.php
@@ -55,9 +55,9 @@ class Article extends Frontend
                 'article_type_options'=>frelink_article_type_options(true),
             ]
         );
-        $this->TDK(($articleType !== 'all' ? frelink_article_type_label($articleType) . ' - ' : '') . '知识内容库');
-		return $this->fetch();
-	}
+        $this->TDK(($articleType !== 'all' ? frelink_article_type_label($articleType) . ' - ' : '') . L('知识内容库'));
+        return $this->fetch();
+    }
 
 	/**
 	 * 发表文章